Refactor sidebar components to utilize bespoke Frame for improved layout and structure
- Replaced standard div elements with the new `Frame` component in various sidebar components, including `VMObjectsSidebar`, `VMNode`, and `VMTopbar`, to enhance styling consistency.
- Introduced a new `RowSelectionTarget` component to manage selection data attributes for sidebar rows, improving the handling of selection states.
- Updated imports to reflect the new bespoke structure, ensuring better organization and maintainability of the codebase.

These changes enhance the overall user experience and maintainability of the editor's sidebar components.
