+++ packages/editor/app/sidebars/objects/VMBoard.tsx
@@ -13,7 +13,7 @@ import { useRowBoard } from "./hooks/use-row-board"
 import { getComponentKey } from "@lib/workspace/workspace-accessors"
 import { ItemNode } from "@seldon/components/elements/ItemNode"
 import { FramerExpandable } from "@app/sidebars/FramerExpandable"
-import { RowSelectionTarget } from "./RowSelectionTarget"
+import { RowSelectionTarget } from "./RowSelectionTarget.bespoke"
 import { VMNode } from "./VMNode"
 import { VMResourceEntry } from "./VMResourceEntry"
 import { getBoardResourceRowConfig } from "./helpers/resource-row-config"
