+++ packages/editor/app/sidebars/objects/VMResourceEntry.tsx
@@ -14,7 +14,7 @@ import { useResourceEntryRow } from "./hooks/use-resource-entry-row"
 import { useRowClick } from "./hooks/use-row-click"
 import { ItemNode } from "@seldon/components/elements/ItemNode"
 import { IconProps } from "@seldon/components/primitives/Icon"
-import { RowSelectionTarget } from "./RowSelectionTarget"
+import { RowSelectionTarget } from "./RowSelectionTarget.bespoke"
 import type { ResourceRowConfig } from "./helpers/resource-row-config"
 
 type VMResourceEntryProps = {
